Warning-clean build + clippy, dead code removed, DPI paths audited (correct;
single-monitor limitation documented), test suite honestly green, ROADMAP
reconciled through phase 16. Completes the Phase 0 P8 polish capstone.

+++ glint/src-tauri/src/capture/cursor.rs
@@ -52,7 +52,7 @@ pub fn composite_cursor(rgba: &mut [u8], width: u32, height: u32, origin_x: i32,
                 biHeight: -CURSOR_SIZE, // negative = top-down
                 biPlanes: 1,
                 biBitCount: 32,
-                biCompression: BI_RGB.0 as u32,
+                biCompression: BI_RGB.0,
                 ..Default::default()
             },
             ..Default::default()
